+++ TongParty/TongParty/DDTabbarViewController.m
@@ -53,7 +53,7 @@
 - (void)didSelectBtnWithBtnTag:(NSNotification *)notification {
     UIButton *btn = (UIButton *)notification.object;
     
-    NSLog(@"-----------%ld",btn.tag);
+    NSLog(@"-----------%ld",(long)btn.tag);
     switch (btn.tag) {
         case 0:{
             DDNavViewController   *nav = [[DDNavViewController alloc]initWithRootViewController:[LSCreateDeskVC new]];
@@ -74,12 +74,12 @@
 }
 //是否允许切换
 - (BOOL)tabBar:(CYTabBar *)tabBar willSelectIndex:(NSInteger)index{
-    NSLog(@"将要切换到---> %ld",index);
+    NSLog(@"将要切换到---> %ld",(long)index);
     return YES;
 }
 //通知切换的下标
 - (void)tabBar:(CYTabBar *)tabBar didSelectIndex:(NSInteger)index{
-    NSLog(@"切换到---> %ld",index);
+    NSLog(@"切换到---> %ld",(long)index);
 }
 
 - (void)didReceiveMemoryWarning {
